changed return value for digit check to double and changed input to double.

+++ com/sample/calculatorapp/CalculatorApp.java
@@ -64,7 +64,7 @@ public class CalculatorApp {
         return selection;
     }
 
-    public int checkUserDigits (int digitsToCheck)
+    public double checkUserDigits (double digitsToCheck)
     {
 
     }
